drivers: misc: adi-axi-tdd: Add new TDD engine driver
This patch introduces the driver for the new ADI TDD engine HDL.
The generic TDD controller is in essence a waveform generator
capable of addressing RF applications which require Time Division
Duplexing, as well as controlling other modules of general
applications through its dedicated 32 channel outputs.

The reason of creating the generic TDD controller was to reduce
the naming confusion around the existing repurposed TDD core
built for AD9361, as well as expanding its number of output
channels for systems which require more than six controlling signals.

The interface is sysfs based and follows the IIO specification, even
tough the device is not actually registered as an IIO device, and
requires a translation layer like iio_fakedev or a modified iiod to be
visible to an iio client.

Signed-off-by: David Winter <david.winter@analog.com>
Signed-off-by: Eliza Balas <eliza.balas@analog.com>